Fixed search bar wrong layout appearance
The search bar’s trailing edge was exceed the trailing edge of view
previously.
After some research of programmatically adding the constraint of search
bar, the solution still not be approached, because of search
controller’s relayout , mapped to whole view , not allow hierarchical
constrained subview.
Despite the unsolved result , I still leave the codes to show how to
add constraint programmatically.
Finally, the exceed issue is solved by adding the sizeToFit() function
after view has completely loaded , which is in the viewDidAppear
function of UIView